refactor CreateSchedule.jsx by creating hooks that it can call to do functionality and add JSDoc comments to some of the hooks that did not have them
