+++ frontend/src/functionality/Calendar/CreateSchedule.jsx
@@ -1,103 +1,34 @@
 import { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
-import { api } from '../../api.js';
 import TimeBlockForm from '../../components/TimeBlockForm.jsx';
 import GeneratorForm from '../../components/GeneratorForm.jsx';
-import generateSchedule from '../../utils/Api/generateSchedule.js';
+import { useCreateTimeBlock } from '../../utils/Hooks/useCreateTimeBlock.js';
+import { useGenerateSchedule } from '../../utils/Hooks/useGenerateSchedule.js';
 
 const TABS = [
-    { id: 'time-block', label: 'Time Block' },
+    { id: 'time-block', label: 'Task' },
     { id: 'generate', label: 'Generate' }
 ];
 
 /**
- * CreateSchedule component with tabs for manual time block creation or schedule generation.
- * @returns {JSX.Element}
+ * CreateSchedule component with tabs for manual task creation or schedule generation.
+ * @returns {JSX.Element} The create schedule page.
  */
 function CreateSchedule() {
-    const navigate = useNavigate();
-    const [serverErrors, setServerErrors] = useState({});
-    const [success, setSuccess] = useState('');
-    const [loading, setLoading] = useState(false);
     const [activeTab, setActiveTab] = useState('time-block');
 
-    async function handleCreate(dataList) {
-        if (loading) return;
+    const {
+        handleCreate,
+        loading: createLoading,
+        serverErrors: createErrors,
+        setServerErrors: clearCreateErrors
+    } = useCreateTimeBlock();
 
-        setServerErrors([]);
-        setLoading(true);
-
-        const errors = [];
-        let allSuccess = true;
-        let createdBlockId = null;
-
-        for (const data of dataList) {
-            // ensure previousely allowed to be "" for start_time and end_time are now null for error handling
-            const cleanedData = {
-                ...data,
-                start_time: data.start_time === '' ? null : data.start_time,
-                end_time: data.end_time === '' ? null : data.end_time
-            };
-
-            try {
-                const res = await api.post('/api/time-blocks/', cleanedData);
-                errors.push({});
-                if (!createdBlockId) createdBlockId = res.data.id;
-            } catch (err) {
-                errors.push(err.response?.data || {});
-                allSuccess = false;
-            }
-        }
-
-        setServerErrors(errors);
-        setLoading(false);
-
-        if (allSuccess)
-            navigate('/successful-time-block', {
-                state: { id: createdBlockId }
-            });
-    }
-
-    /**
-     * Generate schedule via API, store result in session, and navigate to preview.
-     * @param {object} data
-     */
-    async function handleGenerate(data) {
-        if (loading) return;
-
-        setServerErrors({});
-        setLoading(true);
-
-        let allSuccess = true;
-        let response = null;
-
-        try {
-            response = await generateSchedule(data);
-
-            const events = response.data?.events || [];
-
-            if (events.length === 0) {
-                setServerErrors({
-                    general: [
-                        'No feasible schedule could be generated with the given constraints.'
-                    ]
-                });
-                return;
-            }
-            setServerErrors({});
-            sessionStorage.setItem(
-                'generatedSchedule',
-                JSON.stringify(response.data)
-            );
-            navigate('/preview-calendar');
-        } catch (err) {
-
-            setServerErrors(err.response?.data || {});
-            allSuccess = false;
-        } finally {
-            setLoading(false);
-        }
-    }
+    const {
+        handleGenerate,
+        loading: generateLoading,
+        serverErrors: generateErrors,
+        setServerErrors: clearGenerateErrors
+    } = useGenerateSchedule();
 
     return (
         <div>
@@ -105,7 +36,6 @@ function CreateSchedule() {
                 <div className="time-block-form-card">
                     <h2>Create Schedule</h2>
 
-                    {/* Tab Bar */}
                     <div className="tab-bar">
                         {TABS.map((tab) => (
                             <button
@@ -118,15 +48,13 @@ function CreateSchedule() {
                         ))}
                     </div>
 
-                    {/* Tab Content */}
-                    {success && <p>{success}</p>}
 
                     {activeTab === 'time-block' && (
                         <TimeBlockForm
                             onSubmit={handleCreate}
-                            loading={loading}
-                            serverErrors={serverErrors}
-                            clearErrors={() => setServerErrors([])}
+                            loading={createLoading}
+                            serverErrors={createErrors}
+                            clearErrors={() => clearCreateErrors([])}
                         />
                     )}
 
@@ -134,9 +62,9 @@ function CreateSchedule() {
                         <div className="tab-panel">
                             <GeneratorForm
                                 onSubmit={handleGenerate}
-                                loading={loading}
-                                serverErrors={serverErrors}
-                                clearErrors={() => setServerErrors({})}
+                                loading={generateLoading}
+                                serverErrors={generateErrors}
+                                clearErrors={() => clearGenerateErrors({})}
                             />
                         </div>
                     )}
