+++ frontend/src/utils/Hooks/useEditTimeBlockNavigation.js
@@ -13,12 +13,20 @@ import { useNavigate } from 'react-router-dom';
 function useEditTimeBlockNavigation(id) {
     const navigate = useNavigate();
 
+    /**
+     * Navigates to the success page after a time block has been edited.
+     * @returns {void}
+     */
     function goSuccess() {
         navigate('/successful-time-block', {
             state: { id, action: 'edited' }
         });
     }
 
+    /**
+     * Navigates back to the calendar page, discarding any changes.
+     * @returns {void}
+     */
     function goCancel() {
         navigate('/calendar');
     }
